+++ R/latex.R
@@ -61,15 +61,21 @@ to_latex.huxtable <- function (ht, tabular_only = FALSE, ...){
     c(sprintf("\\resizebox*{!}{%s}{", height), "}")
   }
 
+  table_env <- table_environment(ht)
   table_env <- switch(position(ht),
           "wrapleft"  = c("\\begin{wraptable}{l}{%s}", "\\end{wraptable}"),
           "wrapright" = c("\\begin{wraptable}{r}{%s}", "\\end{wraptable}"),
-          c(sprintf("\\begin{table}[%s]", latex_float(ht)), "\\end{table}")
+          c(
+            sprintf("\\begin{%s}[%s]", table_env, latex_float(ht)),
+            sprintf("\\end{%s}", table_env)
+          )
         )
   # no-op except for wraptable:
   wraptable_width <- latex_table_width(ht)
   if (is.na(wraptable_width)) wraptable_width <- "0.25\\textwidth"
-  table_env[1] <- sprintf(table_env[1], wraptable_width)
+  if (position(ht) %in% c("wrapleft", "wrapright")) {
+    table_env[1] <- sprintf(table_env[1], wraptable_width)
+  }
   table_env <- paste0("\n", table_env, "\n")
 
   cap <- build_latex_caption(ht)